Extract trusted policy functions and move script to external
